include/gRPC/python/requirements.txt
matplotlib
numpy
scipy
jupyter

pyqt5==5.14.2

grpcio
grpcio-tools